+++ jax/lax_linalg.py
@@ -21,6 +21,7 @@ import numpy as onp
 from jax.numpy import lax_numpy as np
 from jax import core
 from jax import lax
+from jax import lax_control_flow
 from jax import ad_util
 from jax.interpreters import xla
 from jax.interpreters import ad
@@ -376,8 +377,8 @@ def lu_pivots_to_permutation(swaps, k):
 
   n, = np.shape(swaps)
   permutation = np.arange(k)
-  _, permutation = lax.fori_loop(onp.array(0, onp.int32), onp.array(n, onp.int32),
-                                 body_fn, (swaps, permutation))
+  _, permutation = lax_control_flow.fori_loop(
+      onp.array(0, onp.int32), onp.array(n, onp.int32), body_fn, (swaps, permutation))
   return permutation
