+++ tests/lax_control_flow_test.py
@@ -1,0 +1,410 @@
+# Copyright 2018 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from functools import partial
+
+from absl.testing import absltest
+
+import numpy as onp
+import numpy.random as npr
+
+from jax import api
+from jax import lax
+from jax import lax_control_flow
+from jax import test_util as jtu
+
+
+class LaxControlFlowTest(jtu.JaxTestCase):
+
+  def testWhileWithTuple(self):
+    limit = 10
+
+    def loop_cond(state):
+      pos, _ = state
+      return lax.lt(pos, limit)
+
+    def loop_body(state):
+      pos, count = state
+      return (lax.add(pos, 1), lax.add(count, 1))
+
+    def loop(init):
+      result = lax_control_flow.while_loop(loop_cond, loop_body, (init, 0))
+      _, count = result
+      return count
+
+    cloop = api.jit(loop)
+
+    self.assertEqual(loop(2), limit - 2)
+    self.assertEqual(cloop(2), limit - 2)
+    self.assertEqual(cloop(2), limit - 2)
+    self.assertEqual(cloop(3), limit - 3)
+
+  def testNestedWhile(self):
+
+    def outer_loop(num):  # pylint: disable=missing-docstring
+      def cond_fun(state):
+        num, i, _ = state
+        return lax.lt(i, num)
+
+      def body_fun(state):
+        num, i, count = state
+        return (num, lax.add(i, 1), inner_loop(i, count))
+
+      init_val = (num, 0, 0)
+      _, i, count = lax_control_flow.while_loop(cond_fun, body_fun, init_val)
+      return (i, count)
+
+    def inner_loop(i, count):  # pylint: disable=missing-docstring
+      def cond_fun(state):
+        i, j, _ = state
+        return lax.le(j, i)
+
+      def body_fun(state):
+        i, j, count = state
+        return (i, lax.add(j, 1), lax.add(count, 1))
+
+      init_val = (i, 0, count)
+      _, _, count = lax_control_flow.while_loop(cond_fun, body_fun, init_val)
+      return count
+
+    cloop = api.jit(outer_loop)
+
+    self.assertEqual(outer_loop(3), (3, 6))
+    self.assertEqual(cloop(3), (3, 6))
+    self.assertEqual(cloop(3), (3, 6))
+    self.assertEqual(cloop(2), (2, 3))
+    self.assertEqual(cloop(4), (4, 10))
+
+  def testWhileWithClosure(self):
+
+    def loop(init, local_limit, inc):
+
+      def loop_cond(state):
+        pos, _ = state
+        return lax.lt(pos, local_limit)
+
+      def loop_body(state):
+        effect[0] = True
+        pos, count = state
+        return (lax.add(pos, 1), lax.add(count, inc))
+
+      result = lax_control_flow.while_loop(loop_cond, loop_body, (init, 0))
+      _, count = result
+      return count
+
+    cloop = api.jit(loop)
+
+    limit = 10
+    effect = [False]
+    self.assertEqual(loop(2, limit, 1), limit - 2)
+    assert effect[0]
+    effect[0] = False
+    self.assertEqual(cloop(2, limit, 1), limit - 2)
+    assert effect[0]
+    effect[0] = False
+    self.assertEqual(cloop(2, limit, 1), limit - 2)
+    self.assertEqual(cloop(3, limit, 1), limit - 3)
+    assert not effect[0]
+
+  def testWhileWithClosureJit(self):
+
+    def loop(init, local_limit, inc):
+
+      def loop_cond(state):
+        pos, _ = state
+        return lax.lt(pos, local_limit)
+
+      def loop_body(state):
+        effect[0] = True
+        pos, count = state
+        f = lambda pos, inc: (lax.add(pos, 1), lax.add(count, inc))
+        return api.jit(f)(pos, inc)
+
+      result = lax_control_flow.while_loop(loop_cond, loop_body, (init, 0))
+      _, count = result
+      return count
+
+    cloop = api.jit(loop)
+
+    limit = 10
+    effect = [False]
+    self.assertEqual(loop(2, limit, 1), limit - 2)
+    assert effect[0]
+    effect[0] = False
+    self.assertEqual(cloop(2, limit, 1), limit - 2)
+    assert effect[0]
+    effect[0] = False
+    self.assertEqual(cloop(2, limit, 1), limit - 2)
+    self.assertEqual(cloop(3, limit, 1), limit - 3)
+    assert not effect[0]
+
+  def testNestedWhileWithDynamicUpdateSlice(self):
+    num = 5
+
+    def update_entry(arr, val, i, j):
+      val = lax.reshape(val, [1, 1])
+      return lax.dynamic_update_slice(arr, val, (i, j))
+
+    def outer_loop(arr):  # pylint: disable=missing-docstring
+
+      def cond_fun(state):
+        i, num, _, _ = state
+        return lax.lt(i, num)
+
+      def body_fun(state):
+        i, num, arr, out = state
+        return (lax.add(i, 1), num, arr, inner_loop(i, arr, out))
+
+      out = onp.zeros(arr.shape, dtype=arr.dtype)
+      init_val = (0, num, arr, out)
+      _, _, _, out = lax_control_flow.while_loop(cond_fun, body_fun, init_val)
+      return out
+
+    def inner_loop(i, arr, out):  # pylint: disable=missing-docstring
+
+      def cond_fun(state):
+        i, j, _, _ = state
+        return lax.le(j, i)
+
+      def body_fun(state):
+        i, j, arr, out = state
+        arr_i = lax.dynamic_index_in_dim(arr, i, 0, False)
+        arr_i_j = lax.dynamic_index_in_dim(arr_i, j, 0, False)
+        out = update_entry(out, arr_i_j, i, j)
+        return (i, lax.add(j, 1), arr, out)
+
+      init_val = (i, 0, arr, out)
+      _, _, _, out = lax_control_flow.while_loop(cond_fun, body_fun, init_val)
+      return out
+
+    cloop = api.jit(outer_loop)
+    arr = npr.RandomState(0).randn(5, 5)
+    self.assertAllClose(outer_loop(arr), onp.tril(arr), check_dtypes=False)
+    self.assertAllClose(cloop(arr), onp.tril(arr), check_dtypes=False)
+    self.assertAllClose(cloop(arr), onp.tril(arr), check_dtypes=False)
+
+  def testLoopWithConjunctionCondition(self):
+    def sum_first_n(arr, num):  # pylint: disable=missing-docstring
+      def cond_fun(state):
+        arr, num, i, _ = state
+        return lax.bitwise_and(lax.lt(i, num), lax.lt(i, arr.shape[0]))
+
+      def body_fun(state):
+        arr, num, i, total = state
+        arr_i = lax.dynamic_index_in_dim(arr, i, 0, False)
+        return (arr, num, lax.add(i, 1), lax.add(total, arr_i))
+
+      init_val = (arr, num, 0, 0.)
+      _, _, _, total = lax_control_flow.while_loop(cond_fun, body_fun, init_val)
+      return total
+
+    cfun = api.jit(sum_first_n)
+    x = npr.RandomState(0).randn(10)
+
+    for num in [0, 5, 10, 15]:
+      self.assertAllClose(sum_first_n(x, num), onp.sum(x[:num]),
+                          check_dtypes=False)
+      self.assertAllClose(cfun(x, num), onp.sum(x[:num]), check_dtypes=False)
+      self.assertAllClose(cfun(x, num), onp.sum(x[:num]), check_dtypes=False)
+
+  def testForiLoopBasic(self):
+    def count(num):
+      def body_fun(i, tot):
+        return lax.add(tot, i)
+      return lax_control_flow.fori_loop(0, num, body_fun, 0)
+
+    cfun = api.jit(count)
+
+    self.assertEqual(count(2), 1)
+    self.assertEqual(count(2), cfun(2))
+    self.assertEqual(count(3), 3)
+    self.assertEqual(count(3), cfun(3))
+    self.assertEqual(count(4), 6)
+    self.assertEqual(count(4), cfun(4))
+
+  def testForiLoopClosure(self):
+    def count(num):
+      def body_fun(i, tot):
+        return lax.add(num, lax.add(tot, i))
+      return lax_control_flow.fori_loop(0, num, body_fun, 0)
+
+    cfun = api.jit(count)
+
+    self.assertEqual(count(2), 1 + 2**2)
+    self.assertEqual(count(2), cfun(2))
+    self.assertEqual(count(3), 3 + 3**2)
+    self.assertEqual(count(3), cfun(3))
+    self.assertEqual(count(4), 6 + 4**2)
+    self.assertEqual(count(4), cfun(4))
+
+  def testForiLoopTupleState(self):
+    def sum_first_n(arr, num):
+      def body_fun(i, state):
+        arr, total = state
+        arr_i = lax.dynamic_index_in_dim(arr, i, 0, False)
+        return (arr, lax.add(total, arr_i))
+
+      init_val = (arr, 0.)
+      _, total = lax_control_flow.fori_loop(0, lax.min(arr.shape[0], num), body_fun,
+                               init_val)
+      return total
+
+    cfun = api.jit(sum_first_n)
+    x = npr.RandomState(0).randn(10)
+
+    for num in [0, 5, 10, 15]:
+      self.assertAllClose(sum_first_n(x, num), onp.sum(x[:num]),
+                          check_dtypes=False)
+      self.assertAllClose(cfun(x, num), onp.sum(x[:num]), check_dtypes=False)
+      self.assertAllClose(cfun(x, num), onp.sum(x[:num]), check_dtypes=False)
+
+  def testForiLoopDictState(self):
+    def sum_first_n(arr, num):
+      def body_fun(i, state):
+        arr, total = state['arr'], state['total']
+        arr_i = lax.dynamic_index_in_dim(arr, i, 0, False)
+        return {'arr': arr, 'total': lax.add(total, arr_i)}
+
+      init_val = {'arr': arr, 'total': 0.}
+      out_val = lax_control_flow.fori_loop(0, lax.min(arr.shape[0], num), body_fun, init_val)
+      return out_val['total']
+
+    cfun = api.jit(sum_first_n)
+    x = npr.RandomState(0).randn(10)
+
+    for num in [0, 5, 10, 15]:
+      self.assertAllClose(sum_first_n(x, num), onp.sum(x[:num]),
+                          check_dtypes=False)
+      self.assertAllClose(cfun(x, num), onp.sum(x[:num]), check_dtypes=False)
+      self.assertAllClose(cfun(x, num), onp.sum(x[:num]), check_dtypes=False)
+
+  def testForiLoopEmptyTupleInState(self):
+    def sum_first_n(arr, num):
+      def body_fun(i, state):
+        arr, total, _ = state
+        arr_i = lax.dynamic_index_in_dim(arr, i, 0, False)
+        return (arr, lax.add(total, arr_i), ())
+
+      init_val = (arr, 0., ())
+      _, tot, _ = lax_control_flow.fori_loop(0, lax.min(arr.shape[0], num), body_fun, init_val)
+      return tot
+
+    cfun = api.jit(sum_first_n)
+    x = npr.RandomState(0).randn(10)
+
+    for num in [0, 5, 10, 15]:
+      self.assertAllClose(sum_first_n(x, num), onp.sum(x[:num]),
+                          check_dtypes=False)
+      self.assertAllClose(cfun(x, num), onp.sum(x[:num]), check_dtypes=False)
+      self.assertAllClose(cfun(x, num), onp.sum(x[:num]), check_dtypes=False)
+
+  def testCond(self):
+    def fun(x):
+      if x < 3:
+        return (x, x)
+      else:
+        y = lax.mul(2, x)
+        return y, lax.mul(2, y)
+
+    @api.jit
+    def cfun(x):
+      def false_fun(x):
+        y = lax.mul(2, x)
+        return y, lax.mul(2, y)
+      return lax_control_flow.cond(lax.lt(x, 3), x, lambda x: (x, x), x, false_fun)
+
+    self.assertEqual(fun(0), cfun(0))
+    self.assertEqual(fun(0), (0, 0))
+    self.assertEqual(fun(1), cfun(1))
+    self.assertEqual(fun(1), (1, 1))
+    self.assertEqual(fun(2), cfun(2))
+    self.assertEqual(fun(2), (2, 2))
+    self.assertEqual(fun(3), cfun(3))
+    self.assertEqual(fun(3), (6, 12))
+    self.assertEqual(fun(4), cfun(4))
+    self.assertEqual(fun(4), (8, 16))
+
+  def testNestedCond(self):
+    def fun(x):
+      if x < 2:
+        return lax.mul(2, x)
+      else:
+        if x < 5:
+          return lax.mul(3, x)
+        else:
+          return lax.mul(4, x)
+
+    @api.jit
+    def cfun(x):
+      return lax_control_flow.cond(
+          lax.lt(x, 2),
+          x, lambda x: lax.mul(2, x),
+          x, lambda x: lax_control_flow.cond(lax.lt(x, 5),
+                                x, lambda x: lax.mul(3, x),
+                                4, lambda y: lax.mul(y, x)))
+
+    self.assertEqual(cfun(1), 2)
+    self.assertEqual(cfun(3), 9)
+    self.assertEqual(cfun(6), 24)
+    self.assertEqual(cfun(1), fun(1))
+    self.assertEqual(cfun(3), fun(3))
+    self.assertEqual(cfun(6), fun(6))
+
+  def testCondOneBranchConstant(self):
+    def fun(x):
+      if x < 3:
+        return 5.
+      else:
+        return x
+
+    @api.jit
+    def cfun(x):
+      return lax_control_flow.cond(lax.lt(x, 3), x, lambda x: 5, x, lambda x: x)
+
+    self.assertEqual(fun(0), cfun(0))
+    self.assertEqual(cfun(0), 5)
+    self.assertEqual(fun(4), cfun(4))
+    self.assertEqual(cfun(4), 4)
+
+  def testCondOneBranchConstantTuple(self):
+    def fun(x):
+      if x < 3:
+        return (1., 2., 3.)
+      else:
+        return (x, 2., 4.)
+
+    @api.jit
+    def cfun(x):
+      return lax_control_flow.cond(lax.lt(x, 3),
+                      x, lambda x: (1, 2., 3.),
+                      x, lambda x: (x, 2., 4.))
+
+    self.assertEqual(fun(0), cfun(0))
+    self.assertEqual(cfun(0), (1, 2., 3.))
+    self.assertEqual(fun(4), cfun(4))
+    self.assertEqual(cfun(4), (4, 2., 4.))
+
+  def testIssue514(self):
+    # just check this doesn't crash
+    lax_control_flow.cond(True,
+            (0, 0), lambda x: (x[0], 0),
+            (1, 1), lambda x: x)
+
+
+if __name__ == '__main__':
+  absltest.main()
